+++ dbms/src/Functions/URL/extractURLParameter.cpp
@@ -1,5 +1,5 @@
 #include <Functions/FunctionFactory.h>
-#include <Functions/FunctionsStringSearch.h>
+#include <Functions/FunctionsStringSearchToString.h>
 #include <common/find_symbols.h>
 
 namespace DB
