+++ dbms/src/Functions/multiSearchAllPositions.cpp
@@ -1,0 +1,23 @@
+#include "FunctionsMultiStringPosition.h"
+#include "FunctionFactory.h"
+#include "MultiSearchAllPositionsImpl.h"
+#include "PositionImpl.h"
+
+
+namespace DB
+{
+
+struct NameMultiSearchAllPositions
+{
+    static constexpr auto name = "multiSearchAllPositions";
+};
+
+using FunctionMultiSearchAllPositions
+    = FunctionsMultiStringPosition<MultiSearchAllPositionsImpl<PositionCaseSensitiveASCII>, NameMultiSearchAllPositions>;
+
+void registerFunctionMultiSearchAllPositions(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionMultiSearchAllPositions>();
+}
+
+}
