+++ dbms/src/Functions/multiSearchAnyCaseInsensitiveUTF8.cpp
@@ -1,0 +1,23 @@
+#include "FunctionsMultiStringSearch.h"
+#include "FunctionFactory.h"
+#include "MultiSearchImpl.h"
+#include "PositionImpl.h"
+
+
+namespace DB
+{
+
+struct NameMultiSearchAnyCaseInsensitiveUTF8
+{
+    static constexpr auto name = "multiSearchAnyCaseInsensitiveUTF8";
+};
+
+using FunctionMultiSearchCaseInsensitiveUTF8
+    = FunctionsMultiStringSearch<MultiSearchImpl<PositionCaseInsensitiveUTF8>, NameMultiSearchAnyCaseInsensitiveUTF8>;
+
+void registerFunctionMultiSearchAnyCaseInsensitiveUTF8(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionMultiSearchCaseInsensitiveUTF8>();
+}
+
+}
