+++ dbms/src/Functions/multiSearchFirstIndexCaseInsensitive.cpp
@@ -1,0 +1,23 @@
+#include "FunctionsMultiStringSearch.h"
+#include "FunctionFactory.h"
+#include "MultiSearchFirstIndexImpl.h"
+#include "PositionImpl.h"
+
+
+namespace DB
+{
+
+struct NameMultiSearchFirstIndexCaseInsensitive
+{
+    static constexpr auto name = "multiSearchFirstIndexCaseInsensitive";
+};
+
+using FunctionMultiSearchFirstIndexCaseInsensitive
+    = FunctionsMultiStringSearch<MultiSearchFirstIndexImpl<PositionCaseInsensitiveASCII>, NameMultiSearchFirstIndexCaseInsensitive>;
+
+void registerFunctionMultiSearchFirstIndexCaseInsensitive(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionMultiSearchFirstIndexCaseInsensitive>();
+}
+
+}
