+++ dbms/src/Functions/multiSearchFirstPositionCaseInsensitive.cpp
@@ -1,0 +1,23 @@
+#include "FunctionsMultiStringSearch.h"
+#include "FunctionFactory.h"
+#include "MultiSearchFirstPositionImpl.h"
+#include "PositionImpl.h"
+
+
+namespace DB
+{
+
+struct NameMultiSearchFirstPositionCaseInsensitive
+{
+    static constexpr auto name = "multiSearchFirstPositionCaseInsensitive";
+};
+
+using FunctionMultiSearchFirstPositionCaseInsensitive
+    = FunctionsMultiStringSearch<MultiSearchFirstPositionImpl<PositionCaseInsensitiveASCII>, NameMultiSearchFirstPositionCaseInsensitive>;
+
+void registerFunctionMultiSearchFirstPositionCaseInsensitive(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionMultiSearchFirstPositionCaseInsensitive>();
+}
+
+}
